Worker/requirements.txt
aiokafka
python-dotenv
aiosmtplib
cassandra-driver
email_validator